// web/src/features/access-tokens/data/schema.ts
//
// Copyright (c) 2025 rustmailer.com (https://rustmailer.com)
//
// This file is part of the Bichon Email Archiving Project
//
// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU Affero General Public License as published by
// the Free Software Foundation, either version 3 of the License, or
// (at your option) any later version.
//
// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU Affero General Public License for more details.
//
// You should have received a copy of the GNU Affero General Public License
// along with this program.  If not, see <http://www.gnu.org/licenses/>.


interface AccountInfo {
  id: number;
  email: string;
}

interface RateLimit {
  quota: number;
  interval: number;
}

interface AccessControl {
  ip_whitelist?: string[];
  rate_limit?: RateLimit;
}

interface AccessToken {
  token: string;
  accounts: AccountInfo[];
  created_at: number;
  updated_at: number;
  description?: string;
  last_access_at: number;
  acl?: AccessControl;
}

export type { AccessToken, AccountInfo, AccessControl, RateLimit };